backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5.0.0  # Pin to compatible version with passlib 1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9  # Required for PostgreSQL (Neon)
asyncpg>=0.29.0  # Required for async PostgreSQL connections
pytz>=2024.1  # Timezone support for attendance calculations
reportlab>=4.0.0  # PDF generation for reports
apscheduler>=3.10.0  # Background job scheduler
slowapi>=0.1.9  # API rate limiting
openpyxl>=3.1.0  # Excel generation for reports
redis>=5.0.0  # Redis client for caching, rate limiting, job queue
rq>=1.16.0  # Redis Queue for reliable background jobs

